requirements.txt
Django==6.0.4
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
gunicorn
whitenoise[brotli]
